Feature/post&message api test

+++ src/main/java/com/SafeNet/Backend/domain/post/service/PostService.java
@@ -3,6 +3,7 @@ package com.SafeNet.Backend.domain.post.service;
 import com.SafeNet.Backend.domain.file.entity.File;
 import com.SafeNet.Backend.domain.file.entity.FileType;
 import com.SafeNet.Backend.domain.file.service.FileStorageService;
+import com.SafeNet.Backend.domain.file.service.S3Service;
 import com.SafeNet.Backend.domain.member.entity.Member;
 import com.SafeNet.Backend.domain.member.repository.MemberRepository;
 import com.SafeNet.Backend.domain.post.entity.Post;
@@ -11,8 +12,6 @@ import com.SafeNet.Backend.domain.post.dto.PostRequestDto;
 import com.SafeNet.Backend.domain.post.dto.PostResponseDto;
 import com.SafeNet.Backend.domain.post.exception.PostException;
 import com.SafeNet.Backend.domain.post.repository.PostRepository;
-import com.SafeNet.Backend.domain.postLike.entity.PostLike;
-import com.SafeNet.Backend.domain.postLike.repository.PostLikeRepository;
 import com.SafeNet.Backend.domain.region.entity.Region;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -32,24 +31,30 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
-    private final PostLikeRepository postLikeRepository;
     private final FileStorageService fileStorageService;
     private final MemberRepository memberRepository;
     private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy년 M월 d일");
+    private final S3Service s3Service;
 
     @Transactional
-    public void createPost(PostRequestDto postRequestDto, MultipartFile receiptImage, MultipartFile productImage, Long memberId) {
+    public void createPost(PostRequestDto postRequestDto, MultipartFile receiptImage, MultipartFile productImage, String email) {
+
+        // 사용자 정보로 Member 객체 조회
+        Member member = memberRepository.findByEmail(email).orElseThrow(() -> new PostException("Member not found with email: " + email, HttpStatus.NOT_FOUND));
+
         try {
             // 문자열로 받은 날짜 데이터를 LocalDate로 파싱
             LocalDate parsedBuyDate = LocalDate.parse(postRequestDto.getBuyDate(), formatter);
 
-            // 사용자 정보로 Member 객체 조회
-            Member member = memberRepository.findById(memberId).orElseThrow();
             Region region = member.getRegion();
 
-            // 파일 처리
-            File receiptFile = fileStorageService.saveFile(receiptImage, FileType.receipt);
-            File productFile = fileStorageService.saveFile(productImage, FileType.product_image);
+            // S3에 파일 업로드 및 URL 반환
+            String receiptImageUrl = s3Service.upload("receiptImage", receiptImage.getOriginalFilename(), receiptImage);
+            String productImageUrl = s3Service.upload("productImage", productImage.getOriginalFilename(), productImage);
+
+            // URL을 통해 파일 엔티티 저장
+            File receiptFile = fileStorageService.saveFile(receiptImageUrl, FileType.receipt);
+            File productFile = fileStorageService.saveFile(productImageUrl, FileType.product_image);
 
             // Post 객체 생성 및 저장
             Post post = Post.builder()
@@ -84,19 +89,27 @@ public class PostService {
                         .title(post.getTitle())
                         .writer(post.getMember().getName())
                         .contents(post.getContents())
-                        .imageUrl(post.getFileList().isEmpty() ? null : post.getFileList().get(1).getFileUrl())
+                        .receiptImageUrl(post.getFileList().isEmpty() ? null : post.getFileList().get(0).getFileUrl())
+                        .productImageUrl(post.getFileList().isEmpty() ? null : post.getFileList().get(1).getFileUrl())
                         .cost(post.getCost())
                         .count(post.getCount())
                         .buyDate(post.getBuyDate().toString())
-                        .category(post.getCategory()).build())
+                        .category(post.getCategory())
+                        .likeCount(post.getPostLikeList().size())
+                        .build())
                 .orElseThrow(() -> new PostException("Post not found with id: " + id, HttpStatus.NOT_FOUND)));
     }
 
     @Transactional
-    public void updatePost(Long id, PostRequestDto postRequestDto, MultipartFile receiptImage, MultipartFile productImage) {
-        try {
-            Post existingPost = postRepository.findById(id).orElseThrow(() -> new PostException("Post not found with id: " + id, HttpStatus.NOT_FOUND));
+    public void updatePost(Long id, PostRequestDto postRequestDto, MultipartFile receiptImage, MultipartFile productImage, String email) {
+        Post existingPost = postRepository.findById(id).orElseThrow(() -> new PostException("Post not found with id: " + id, HttpStatus.NOT_FOUND));
 
+        // 사용자 이메일과 게시물 작성자의 이메일 비교
+        if (!existingPost.getMember().getEmail().equals(email)) {
+            throw new PostException("You do not have permission to update this post", HttpStatus.FORBIDDEN);
+        }
+
+        try {
             if (existingPost.getPostStatus() != PostStatus.거래가능) {
                 throw new PostException("Cannot update post with id: " + id + " because its status is not 'AVAILABLE'.", HttpStatus.BAD_REQUEST);
             }
@@ -105,17 +118,29 @@ public class PostService {
             List<File> fileList = existingPost.getFileList();
 
             if (receiptImage != null && !receiptImage.isEmpty()) {
+                // S3와 File 엔티티에 파일 삭제
+                s3Service.delete(fileList.get(0).getFileUrl());
                 fileStorageService.deleteFile(fileList.get(0));
-                File receiptFile = fileStorageService.saveFile(receiptImage, FileType.receipt);
+
+                // S3와 File 엔티티에 파일 생성
+                String receiptImageUrl = s3Service.upload("receiptImage", receiptImage.getOriginalFilename(), receiptImage);
+                File receiptFile = fileStorageService.saveFile(receiptImageUrl, FileType.receipt);
+
+                // 기존 fileList의 첫 번째 요소를 새로 업로드된 영수증 이미지 파일로 대체
                 fileList.set(0, receiptFile);
             }
-
             if (productImage != null && !productImage.isEmpty()) {
+                // S3와 File 엔티티에 파일 삭제
+                s3Service.delete(fileList.get(1).getFileUrl());
                 fileStorageService.deleteFile(fileList.get(1));
-                File productFile = fileStorageService.saveFile(productImage, FileType.product_image);
+
+                // S3와 File 엔티티에 파일 생성
+                String productImageUrl = s3Service.upload("productImage", productImage.getOriginalFilename(), productImage);
+                File productFile = fileStorageService.saveFile(productImageUrl, FileType.product_image);
+
+                // 기존 fileList의 두 번째 요소를 새로 업로드된 제품 이미지 파일로 대체
                 fileList.set(1, productFile);
             }
-
             existingPost.updatePost(postRequestDto, parsedBuyDate, fileList);
             postRepository.save(existingPost);
         } catch (Exception e) {
@@ -124,37 +149,41 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePost(Long id) {
-        try {
-            Post post = postRepository.findById(id).orElseThrow(() -> new PostException("Post not found with id: " + id, HttpStatus.NOT_FOUND));
+    public void deletePost(Long id, String email) {
+        Post post = postRepository.findById(id).orElseThrow(() -> new PostException("Post not found with id: " + id, HttpStatus.NOT_FOUND));
 
+        // 사용자 이메일과 게시물 작성자의 이메일 비교
+        if (!post.getMember().getEmail().equals(email)) {
+            throw new PostException("You do not have permission to delete this post", HttpStatus.FORBIDDEN);
+        }
+
+        try {
             if (post.getPostStatus() != PostStatus.거래가능) {
                 throw new PostException("Cannot delete post with id: " + id + " because its status is not 'AVAILABLE'.", HttpStatus.BAD_REQUEST);
             }
-
-            post.getFileList().forEach(fileStorageService::deleteFile);
+            // S3에서 파일 삭제
+            post.getFileList().forEach(file -> {
+                s3Service.delete(file.getFileUrl());
+            });
+            // Post와 연관된 File 엔티티는 cascade = CascadeType.ALL로 인해 자동으로 삭제됨
             postRepository.delete(post);
         } catch (Exception e) {
             throw new PostException("Failed to delete post", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    public List<PostResponseDto> getPostsByMemberId(Long memberId) {
-        try {
-            List<Post> posts = postRepository.findByMember_Id(memberId);
-            return posts.stream().map(this::convertToDto).collect(Collectors.toList());
-        } catch (Exception e) {
-            throw new PostException("Failed to retrieve posts by memberId", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+    @Transactional
+    public void updatePostStatusToTrading(Long id) {
+        Post post = postRepository.findById(id).orElseThrow(() -> new PostException("Post not found", HttpStatus.NOT_FOUND));
+        post.setPostStatus(PostStatus.거래중);
+        postRepository.save(post);
     }
 
-    public List<PostResponseDto> getLikedPostsByMemberId(Long memberId) {
-        try {
-            List<PostLike> postLikes = postLikeRepository.findByMember_Id(memberId);
-            return postLikes.stream().map(postLike -> this.convertToDto(postLike.getPost())).collect(Collectors.toList());
-        } catch (Exception e) {
-            throw new PostException("Failed to retrieve liked posts by memberId", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+    @Transactional
+    public void updatePostStatusToCompleted(Long id) {
+        Post post = postRepository.findById(id).orElseThrow(() -> new PostException("Post not found", HttpStatus.NOT_FOUND));
+        post.setPostStatus(PostStatus.거래완료);
+        postRepository.save(post);
     }
 
     private PostResponseDto convertToDto(Post post) {
@@ -162,7 +191,7 @@ public class PostService {
                 .postId(post.getId())
                 .title(post.getTitle())
                 .likeCount(post.getPostLikeList().size())
-                .imageUrl(post.getFileList().isEmpty() ? null : post.getFileList().get(1).getFileUrl())
+                .productImageUrl(post.getFileList().isEmpty() ? null : post.getFileList().get(1).getFileUrl())
                 .count(post.getCount())
                 .cost(post.getCost())
                 .build();
